Feature/purchases merge to main (#33)
* Adds EventTracker and Purchase events (#30)

Adds an ATTNEventTracker. This class is a singleton so it's easily accessible anywhere in the app. This class is used to send events to Attentive.
Adds OCMock library for testing
Updates Example app with a button to send a Purchase

* Sends Purchase events to server (#31)

Hooks up ATTNEventTracker to the ATTNAPI
Updates ATTNAPI to send PurchaseEvents to the server
Adds nullable annotations to some properties

* Adds OrderConfirmed event (#32)

Sends an OrderConfirmed event when a Purchase event is sent